afxdp/cos/admission: relocate inline tests to admission_tests.rs
admission.rs reaches 1,938 LOC with a 1,437-LOC `#[cfg(test)] mod tests` block.
Same playbook as #1080-#1089: extract to a sibling admission_tests.rs.

After: admission.rs drops from 1,938 → 504 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
